finalisation de la version console avec adaptation comme demandée par le prof : le joueur peut tirer dans toutes les directions (float) selon un angle donné (avec la fonction updateTirConsole) puis arrondi en int dans mainconsole pour le deplacement sur cases

+++ src/Direction.h
@@ -2,8 +2,8 @@
 #define DIRECTION_H
 
 struct Direction {
-    int dx;
-    int dy;
+    float dx;
+    float dy;
 };
 
 #endif
